backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
recipe-scrapers>=14.52.0
yt-dlp>=2025.1.26
openai>=1.12.0
beautifulsoup4>=4.12.3
requests>=2.31.0
sqlalchemy>=2.0.25
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic>=2.6.0
httpx>=0.26.0
